Continued implementation of Property classes, and addition of exists() to FileReader and FileWriter for checking whether a file already exists

+++ TBTK/calc/TightBindingLib/include/Dos.h
@@ -17,8 +17,8 @@ public:
 	Dos(double lowerLimit, double upperLimit, int resolution);
 	~Dos();
 
-	int getLowerLimit();
-	int getUpperLimit();
+	double getLowerLimit();
+	double getUpperLimit();
 	int getResolution();
 	const double* getData();
 private:
